NR_REG_SINGLE_RX: support single RX queue mode
When an adaptor is in single RX queue mode, only one RX ring is
actually in nm mode, while all the other ones (including all the TX
ones) are left attached to the host network stack.

To setup the adapter in this mode it is sufficient to set the nm_req
flag field to NR_REG_SINGLE_RX and ring_id equal to the id of the
requested rx queue.

Note that this mode cannot be mixed with other modes: after opening an
adapter in single rx queue mode, it is not possible to reopen it in a
different mode, and viceversa, after that it has been opened in a
different mode than the single rx queue one, this mode cannot be used.