NR_REG_SINGLE_RX: do not initialize host rings in native Netmap mode
When an adapter is in single rx queue mode, only the rings that are
requested to be in Netmap mode should be initialized in the native
Netmap mode (i.e. mapped in the Netmap memory region).

For this reason netmap_reset() has been modified to return NULL if it is
called with a queue that is not supposed to be in Netmap mode.

To avoid races between the queue request and the actual netmap_reset()
call, a second rx_queue bitmap has been introduced, to differentiate a
queue that is actually in Netmap mode between a queue that has been only
requested to be in Netmap mode but is not fully initialized.

Finally everytime a new interface is registered (unregistered), the
adapter callback nm_register() is called to ensure that the requested
ring is put in (removed from) Netmap mode.